requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
ultralytics==8.0.196
# torch and torchvision are installed in Dockerfile with CPU versions
# torch==2.0.1
# torchvision==0.15.2
opencv-python-headless==4.9.0.80
mediapipe==0.10.14
pillow==10.2.0
prometheus-client==0.19.0
ruff==0.1.14
pre-commit>=3.5.0
numpy<2.0.0
aiosqlite==0.19.0
prometheus-fastapi-instrumentator==6.1.0
matplotlib==3.8.2
seaborn==0.13.1

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Frontend
streamlit
requests==2.31.0
websockets
